more restructuring dependency cleaning
