+++ src/components/SendPublicMessage.js
@@ -3,7 +3,7 @@ import React, { useState, useEffect } from "react";
 import { contract, w3 } from "../helpers/Web3Helper";
 
 export default function SendPublicMessage() {
-  const [message, setMessage] = useState();
+  const [message, setMessage] = useState("");
   const [errorMessage, setErrorMessage] = useState();
   const [account, setAccount] = useState();
 
@@ -15,7 +15,7 @@ export default function SendPublicMessage() {
 
   const mySubmitHandler = async (event) => {
     event.preventDefault();
-    if (message === "") {
+    if (!message || message === "") {
       setErrorMessage("Your message cannot be empty");
     } else {
       contract.methods
